+++ src/main/java/chess/dto/game/MoveDTO.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.game;
 
 public final class MoveDTO {
     private String roomId;
