+++ src/main/java/chess/dto/result/ResultDTO.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.result;
 
 public final class ResultDTO {
     private String roomId;
