+++ src/main/java/chess/dto/room/RoomIdDTO.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.room;
 
 public final class RoomIdDTO {
     private String roomId;
