+++ src/main/java/chess/dto/room/RoomNameDTO.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.room;
 
 public final class RoomNameDTO {
     private String name;
